* '임시' of https://github.com/JungEunee/YOGI:
  모임수정시 알람 완료
  지수누나랑 합치기 완료
  재욱 : 모임 댓글 달릴때마다 알람 & 댓글 삭제
  마이페이지 sql변경
  관리자 멤버 리스트 검색 완료 & 모임 개설 스크립트 주석 삭제
  장소 신청/취소 확인창 추가
  장소 삭제 완전히!!!!!!끝!!! (신청되어있는 장소일시 삭제x,댓글달린 장소여도 같이 삭제o)
  장소등록폼에 달력 추가 , 모임수정폼 추가
  관리자 장소 리스트에서 장소 이름 누르면 상세보기로 이동, 만약 장소에 대관신청한 사람이 있으면 수정 불가능하게 하는 기능 추가
  장소 상세페이지에서 주소 클릭시 지도 팝업창 뜨는거 추가, 관리자페이지에서 장소 수정가능하도록 기능 추가
  그룹 사진 등록(개설,수정)시 나오는 null 오류 수정
  장소 등록 - 파일 업로드
  장소 상세보기에서 후기(댓글)  삭제 기능 완료 , (댓글달려있을시 내용만 바뀜, 부모댓글이 모두 삭제되어 내용만 바뀌어있는 상태에서 제일 아래 자식 댓글을 지우면 삭제되있는 위의 부모댓글을까지 같이 삭제)
  재욱 : 대댓+리스트 모임 비활 걸르기+그룹 내용 통합

# Conflicts:
#	YOGI/src/main/java/yogi/group/GroupController.java
#	YOGI/src/main/java/yogi/members/MembersController.java
#	YOGI/src/main/webapp/WEB-INF/views/group/groupDetail.jsp
#	YOGI/src/main/webapp/WEB-INF/views/main/main.jsp

+++ YOGI/src/main/java/yogi/lendplace/LendplaceController.java
@@ -1,5 +1,6 @@
 package yogi.lendplace;
 
+import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
@@ -12,7 +13,10 @@ import org.apache.log4j.Logger;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -61,6 +65,22 @@ public class LendplaceController {
 		return mv;
 	}
 	
+	//장소 수정폼
+		@RequestMapping(value="/lendplace/updateForm")
+		public ModelAndView updateLendplaceForm(CommandMap commandMap) throws Exception{
+			ModelAndView mv = new ModelAndView("/admin/lendplaceUpdateForm");
+			Map<String,Object> map = lendplaceService.selectLendplaceDetail(commandMap.getMap());
+			mv.addObject("map", map);
+			return mv;  
+		}
+	
+	//장소 수정
+	@RequestMapping(value="/lendplace/update")
+	public ModelAndView updateLendplace(CommandMap commandMap) throws Exception{
+	         lendplaceService.updateLendplace(commandMap.getMap());
+	         return new ModelAndView("redirect:/admin/lendplace/list"); 
+	}
+	
 	//장소 신청
 	@RequestMapping(value="/lendplace/apply")
     public ModelAndView placebookInsert(CommandMap commandMap) throws Exception{
@@ -87,11 +107,23 @@ public class LendplaceController {
     //후기 삭제
     @RequestMapping(value="/lendplace/deleteReview")
     public ModelAndView deleteReview(CommandMap commandMap) throws Exception{
-    	lendplaceService.deleteReview(commandMap.getMap());
-    	return new ModelAndView("redirect:/lendplace/detail?L_NO=" + commandMap.getMap().get("L_NO"));
+    	ModelAndView mv = new ModelAndView("redirect:/lendplace/detail?L_NO=" + commandMap.getMap().get("L_NO"));
+    	Map<String,Object> map = lendplaceService.selectReviewChild(commandMap.getMap());
+    	Integer cnt = Integer.parseInt(map.get("CNT").toString());
+    	if ( cnt == 0) {
+    		Map<String,Object> del = lendplaceService.selectDeletedParent(commandMap.getMap());
+    		Map<String,Object> par = lendplaceService.selectParent(commandMap.getMap());
+    		Integer DEL = Integer.parseInt(del.get("DEL").toString());
+    		Integer PAR = Integer.parseInt(par.get("PAR").toString());
+    		
+    		lendplaceService.deleteReview(commandMap.getMap());
+    		if ( DEL == (PAR-1)) {
+				lendplaceService.deleteGroupReview(commandMap.getMap());
+			}
+        	return mv;	
+        }else {
+        	lendplaceService.updateDeleteFlag(commandMap.getMap());
+    		return mv;
+        }
     }
-
-   
-
-
 }
